Moved license file extraction code to CoreLicenseFileService. Moved some test utilities to Core test project to enable testing moved code.
