+++ src/NuGetGallery.Core/Services/CoreLicenseFileService.cs
@@ -7,7 +7,9 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using NuGet.Packaging;
 using NuGet.Services.Entities;
+using NuGetGallery.Packaging;
 
 namespace NuGetGallery
 {
@@ -50,6 +52,36 @@ namespace NuGetGallery
                 : CoreConstants.TextContentType;
 
             return _fileStorageService.SaveFileAsync(_metadata.PackageContentFolderName, fileName, contentType, licenseFile, overwrite: true);
+        }
+
+        public async Task ExtractAndSaveLicenseFileAsync(Package package, Stream packageStream)
+        {
+            if (package == null)
+            {
+                throw new ArgumentNullException(nameof(package));
+            }
+
+            if (packageStream == null)
+            {
+                throw new ArgumentNullException(nameof(packageStream));
+            }
+
+            packageStream.Seek(0, SeekOrigin.Begin);
+            using (var packageArchiveReader = new PackageArchiveReader(packageStream, leaveStreamOpen: true))
+            {
+                var packageMetadata = PackageMetadata.FromNuspecReader(packageArchiveReader.GetNuspecReader(), strict: true);
+                if (packageMetadata.LicenseMetadata == null || packageMetadata.LicenseMetadata.Type != LicenseType.File || string.IsNullOrWhiteSpace(packageMetadata.LicenseMetadata.License))
+                {
+                    throw new InvalidOperationException("No license file specified in the nuspec");
+                }
+
+                var filename = packageMetadata.LicenseMetadata.License;
+                var licenseFileEntry = packageArchiveReader.GetEntry(filename); // throws on non-existent file
+                using (var licenseFileStream = licenseFileEntry.Open())
+                {
+                    await SaveLicenseFileAsync(package, licenseFileStream);
+                }
+            }
         }
 
         public Task<Stream> DownloadLicenseFileAsync(Package package)
